SwahiliBox-OpenDataIoT-June-2015/pi-data-logger/requirements.txt
SqlAlchemy==1.0.5
pyserial==2.7